qa: move `MapperBuilder#warmup` tests to integration tests
